+++ src/app/ApplicationTemplate.Presentation/ViewModels/DadJokes/DadJokesFiltersPageViewModel.cs
@@ -1,26 +1,16 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Reactive.Linq;
-using System.Reactive.Subjects;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 using ApplicationTemplate.Business;
 using Chinook.DynamicMvvm;
 using Chinook.SectionsNavigation;
-using Chinook.StackNavigation;
 
 namespace ApplicationTemplate.Presentation;
 
-public class DadJokesFiltersPageViewModel : ViewModel
+public sealed class DadJokesFiltersPageViewModel : ViewModel
 {
-	public DadJokesFiltersPageViewModel()
-	{
-	}
-
 	public IDynamicCommand HandleCheck => this.GetCommand((string pt) =>
 	{
-		PostTypeFilter = (PostTypes)Enum.Parse(typeof(PostTypes), pt, true);
+		PostTypeFilter = Enum.Parse<PostTypes>(pt, true);
 	});
 
 	public IDynamicCommand FiltersAndNavigate => this.GetCommandFromTask(async ct =>
@@ -31,7 +21,7 @@ public class DadJokesFiltersPageViewModel : ViewModel
 
 	public PostTypes PostTypeFilter
 	{
-		get => this.GetFromTask<PostTypes>(ct => this.GetService<IDadJokesService>().GetAndObservePostTypeFilter().FirstAsync(ct), initialValue: PostTypes.Hot);
+		get => this.GetFromTask(ct => this.GetService<IDadJokesService>().GetAndObservePostTypeFilter().FirstAsync(ct), initialValue: PostTypes.Hot);
 		set => this.Set(value);
 	}
 }